code/requirements.txt
matplotlib
numpy
pandas
pyinstrument
pyroomacoustics
pystoi
pyyaml
resampy
scipy
silero-vad
soundfile
